Change API to return original text
When processing a lot of smaller messages, particularly in the early
stages of development, it is cumbersome to have to keep track of which
text I submitted and what position it is in my data structure.

This change modifies the API to return the input text with the predicted
values. It incurs the cost of another layer of nested dictionaries.